dev/src/backend/requirements.txt
autopep8==2.0.2
blinker==1.6.2
certifi==2021.5.30
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
fpdf2==2.4.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.4.0
Werkzeug==2.3.3
word-search-generator==3.3.0